// Web/web-app/web-app/src/app/modules/reports/models/row-info-aggrigated.ts
export interface RowInfoAggrigated {
  plan: {
    jan: number;
    feb: number;
    mar: number;
    apr: number;
    may: number;
    jun: number;
    jul: number;
    aug: number;
    sep: number;
    oct: number;
    nov: number;
    dec: number;
  },
  actual: {
    jan: number;
    feb: number;
    mar: number;
    apr: number;
    may: number;
    jun: number;
    jul: number;
    aug: number;
    sep: number;
    oct: number;
    nov: number;
    dec: number;
  },
  category: string;
  currency: string;
}